requirements.txt
executorch==1.0.0
torchcodec==0.9.1
torchaudio==2.9.1
pyserial